+++ pages/contact.php
@@ -69,7 +69,7 @@ echo $OUTPUT->header();
 echo $OUTPUT->heading(theme_boost_union_get_staticpage_pagetitle('contact'));
 
 // Output contact page content.
-echo format_text($config->contactcontent);
+echo format_text($config->contactcontent, FORMAT_MOODLE, ['trusted' => true, 'noclean' => true]);
 
 // Finish page.
 echo $OUTPUT->footer();